Rename issues.body to description
Add create issue handler
