+++ bin/gen_server.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 
-swagger generate server -f swagger.yaml -t pkg/ --server-package genserver --model-package messages --exclude-main -A mymove
+mkdir -p pkg/gen
+swagger generate server -f swagger.yaml -t pkg/gen/ --server-package genserver --model-package messages --exclude-main -A mymove
